Add DefaultInterpolatedStringHandler for .NET Framework
Add the code necessary to get minimal allocation string formatting with numbers on .NET Framework.

Accomplished this by porting .NET 6 code used for number ISpanFormattable support. .NET 7 and later started to utilize static abstract interface methods for this, which is not nearly as plausible to port as that feature requires .NET runtime support.